Added config and speech_processing modules
