+++ .gitignore
@@ -1,6 +1,8 @@
 venv
 __pycache__
 .git
+*.log
+.env
 OSR_in_000_0063_8k.wav
 OSR_uk_000_0020_8k.wav
 speech.html
